# docs/swagger.yaml
openapi: 3.0.3
info:
  title: GPTless Actual Intelligence API
  version: 1.0.0
servers:
  - url: https://api.gptless.au
  - url: https://preview.gptless.au
...
security:
  - AccessTokenAuth: []
paths:
  /v1/requester/chats:
    post:
      tags: [Requester side]
      summary: create a new chat with initial request
      description: |
        Creates a new chat and submits the initial request.

      parameters:
        - $ref: "#/components/parameters/AccessToken"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatRequest"

      responses:
        "201":
          description: |
            **Created**
            
            Returns the newly created chatID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

        "400":
          description: |
            **Bad Request**
            
            Missing or invalid chat data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                BAD_REQUEST:
                  summary: BAD_REQUEST
                  value:
                    error: BAD_REQUEST
                    message: "Missing or invalid chat data: missing field: requestText"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to create chats.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
    get:
      tags: [Requester side]
      summary: Lists chats
      description: Provides a list of chats that the requester has made.

      parameters:
        - $ref: "#/components/parameters/AccessToken"

      responses:
        "200":
          description: |
            **OK**
            
            Returns an array of chats.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSummary"
              example:
                - chatID: 12345
                  title: website design
                  category: development
                  status: open
                  tokens: 30
                  createdAt: "2025-08-14T10:30:00Z"
                - chatID: 54321
                  title: poem request
                  category: writing
                  status: closed
                  tokens: 40
                  createdAt: "2025-08-13T09:00:00Z"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/requester/chats/{chatID}:
    parameters:
      - $ref: "#/components/parameters/AccessToken"
      - $ref: "#/components/parameters/ChatId"

    get:
      tags: [Requester side]
      summary: get a chat
      description: Given a valid chatID, returns the full chat.
      responses:
        "200":
          description: |
            **OK**

            Returns the chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/requester/chats/{chatID}/messages:
    post:
      tags: [Requester side]
      summary: send a message to the responder
      description: |
        Sends a message in an existing chat without creating a new paid request, for example providing details or clarification.

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - $ref: "#/components/parameters/ChatId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageRequest"

      responses:
        "201":
          description: |
            **Created**
            
            Returns the newly created message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

        "400":
          description: |
            **Bad Request**
            
            Missing or invalid message data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                BAD_REQUEST:
                  summary: BAD_REQUEST
                  value:
                    error: BAD_REQUEST
                    message: "Missing or invalid message data: missing field: message"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/requester/chats/{chatID}/requests:
    post:
      tags: [Requester side]
      summary: add a new request to an existing chat
      description: |
        Creates a new paid request in an existing chat using the previous message history and context.

      parameters:
        - $ref: "#/components/parameters/AccessToken"
        - $ref: "#/components/parameters/ChatId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewRequestRequest"

      responses:
        "201":
          description: |
            **Created**
            
            Returns the newly created request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Request"

        "400":
          description: |
            **Bad Request**
            
            Missing or invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                BAD_REQUEST:
                  summary: BAD_REQUEST
                  value:
                    error: BAD_REQUEST
                    message: "Missing or invalid request data: missing field: requestText"

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "402":
          description: |
            **Payment Required**
            
            You do not have enough tokens to create a new request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                PAYMENT_REQUIRED:
                  summary: PAYMENT_REQUIRED
                  value:
                    error: PAYMENT_REQUIRED
                    message: You do not have enough tokens to create a new request

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
  /v1/requester/chats/{chatID}/close:
    parameters:
      - $ref: "#/components/parameters/AccessToken"
      - $ref: "#/components/parameters/ChatId"

    post:
      tags: [Requester side]
      summary: mark a chat as closed.
      description: flags a chat as closed.
      responses:
        "200":
          description: |
            **OK**
            
            Chat closed.

        "401":
          description: |
            **Unauthorized**
            
            Missing or invalid API token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                UNAUTHORIZED:
                  summary: UNAUTHORIZED
                  value:
                    error: UNAUTHORIZED
                    message: The API token is missing or invalid. If you do not have an API token register for one through the form on our website

        "403":
          description: |
            **Forbidden**
            
            You do not have access to this chat.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                FORBIDDEN:
                  summary: FORBIDDEN
                  value:
                    error: FORBIDDEN
                    message: You do not have access to this content

        "404":
          description: |
            **Not Found**
            
            ChatID does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                NOT_FOUND:
                  summary: NOT_FOUND
                  value:
                    error: NOT_FOUND
                    message: The requested resource was not found

        "500":
          description: |
            **Internal Server Error**
            
            There was an internal server error. If this persists contact us at contact@gptless.au.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                INTERNAL_SERVER_ERROR:
                  summary: INTERNAL_SERVER_ERROR
                  value:
                    error: INTERNAL_SERVER_ERROR
                    message: An internal server error occurred.
...
components:
  securitySchemes:
    AccessTokenAuth:
      type: apiKey
      in: header
      name: AccessToken
  parameters:
    ChatId:
      name: chatID
      in: path
      required: true
      schema:
        type: string
      description: chatID
    AccessToken:
      name: AccessToken
      in: header
      required: true
      schema:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
      description: AccessToken
  schemas:
    CreateChatRequest:
      type: object
      properties:
        requestText:
          type: string
          example: Create a website for my online store.
        title:
          type: string
          example: website design
        category:
          type: string
          example: writing
        tokensToSpend:
          type: number
          example: 40
      required:
        - requestText
    MessageRequest:
      type: object
      properties:
        message:
          type: string
          example: could you resize the title please.
      required:
        - message
    NewRequestRequest:
      type: object
      properties:
        requestText:
          type: string
          example: Now make a mobile version for it as well.
        tokensToSpend:
          type: number
          example: 70
      required:
        - requestText
...
    ChatSummary:
      type: object
      properties:
        chatID:
          type: string
          example: 12345
        title:
          type: string
          example: poem request
        category:
          type: string
          example: writing
        status:
          type: string
          example: open
        Tokens:
          type: number
          example: 30
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:30:00Z"
    Message:
      type: object
      properties:
        messageID:
          type: string
          example: msg_001
        senderType:
          type: string
          example: requester
        message:
          type: string
          example: Please make the tone more professional.
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:35:00Z"
    Request:
      type: object
      properties:
        requestID:
          type: string
          example: req_001
        requestText:
          type: string
          example: Now make a social media version using the same context.
        status:
          type: string
          example: pending
        tokensSpent:
          type: number
          example: 15
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:40:00Z"
...
    Chat:
      type: object
      properties:
        chatID:
          type: string
          example: 12345
        requesterUsername:
          type: string
          example: Guy
        responderUsername:
          type: string
          nullable: true
          example: John
        status:
          type: string
          example: open
        createdAt:
          type: string
          format: date-time
          example: "2025-08-14T10:30:00Z"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        requests:
          type: array
          items:
            $ref: "#/components/schemas/Request"
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string